InterfaceClass benennt, inventoryManager in eine separate Datei.
