+++ Engine/Entity_Classes/Interface.py
@@ -4,7 +4,7 @@ import Main.generation as generation
 import Engine.Entity_Classes.collectable as collectable
 import Main.settings as settings
 
-class InventorySlot(entities.Entity):
+class Interface(entities.Entity):
     """Klasse, die ein Inventory Slot als Entity definiert. \tpos_x: float für X-Position, \tpos_x: float für Y-Position, \trect: pygame.Rect für Xpos/Ypos/Höhe/Breite, \t scale: Tupel für Breite/Höhe, \tsource: String für Dateipfad"""
     def __init__(self, pos_x, pos_y, rect, scale, source):
         entities.Entity.__init__(self, pos_x, pos_y, rect, "midbottom", scale, source, False, False, True)
